+++ src/hosted-fields/shared/browser-detection.js
@@ -1,6 +1,8 @@
 'use strict';
 
 module.exports = {
+  isIE: require('@braintree/browser-detection/is-ie'),
+  isEdge: require('@braintree/browser-detection/is-edge'),
   isIe9: require('@braintree/browser-detection/is-ie9'),
   isIos: require('@braintree/browser-detection/is-ios'),
   isIosWebview: require('@braintree/browser-detection/is-ios-webview')